5/8 chkpt

+++ droidlet/interpreter/robot/default_behaviors.py
@@ -172,7 +172,7 @@ def get_task_data_for_reexplore(agent):
 def reexplore(agent):
     task_data = get_task_data_for_reexplore(agent)
     
-    if os.path.isdir(task_data['data_path']):
+    if task_data is not None and os.path.isdir(task_data['data_path']):
         logging.info(f"rmtree {task_data['data_path']} ...")
         shutil.rmtree(task_data['data_path'])
